backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
psycopg[binary]==3.2.3
psycopg-pool==3.2.3
pgvector==0.3.6
google-genai==0.3.0
python-dotenv==1.0.1
